+++ Classes/Model/Type/ProductModel.php
@@ -34,6 +34,7 @@ final class ProductModel extends AbstractType
         'gtin13',
         'gtin14',
         'gtin8',
+        'hasEnergyConsumptionDetails',
         'height',
         'identifier',
         'image',