+++ Classes/Model/Type/Researcher.php
@@ -1,0 +1,36 @@
+<?php
+declare(strict_types=1);
+
+/*
+ * This file is part of the "schema" extension for TYPO3 CMS.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ */
+
+namespace Brotkrueml\Schema\Model\Type;
+
+use Brotkrueml\Schema\Core\Model\AbstractType;
+
+/**
+ * Researchers.
+ */
+final class Researcher extends AbstractType
+{
+    protected static $propertyNames = [
+        'additionalType',
+        'alternateName',
+        'audienceType',
+        'description',
+        'disambiguatingDescription',
+        'geographicArea',
+        'identifier',
+        'image',
+        'mainEntityOfPage',
+        'name',
+        'potentialAction',
+        'sameAs',
+        'subjectOf',
+        'url',
+    ];
+}