+++ Classes/Model/Type/Vehicle.php
@@ -40,6 +40,7 @@ final class Vehicle extends AbstractType
         'gtin13',
         'gtin14',
         'gtin8',
+        'hasEnergyConsumptionDetails',
         'height',
         'identifier',
         'image',